move internal mailchain to cmd
