+++ cmd/mailchain/commands/config_ethereum.go
@@ -17,7 +17,7 @@ package commands
 import (
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/config/names"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/setup"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/setup"
 	"github.com/mailchain/mailchain/internal/pkg/encoding"
 	"github.com/spf13/cobra"
 )
