+++ cmd/mailchain/commands/config_sent.go
@@ -17,8 +17,8 @@ package commands
 import (
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/config/names"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/prerun"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/setup"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/prerun"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/setup"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
