+++ cmd/mailchain/config/sent.go
@@ -16,7 +16,7 @@ package config
 
 import (
 	"github.com/mailchain/mailchain/cmd/mailchain/config/names"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/prompts"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/prompts"
 	"github.com/mailchain/mailchain/internal/pkg/stores"
 	"github.com/mailchain/mailchain/internal/pkg/stores/s3"
 	"github.com/pkg/errors"
