+++ cmd/mailchain/internal/commands/prompts/select.go
