+++ cmd/mailchain/internal/commands/setup/network.go
@@ -16,7 +16,7 @@ package setup
 
 import (
 	"github.com/mailchain/mailchain/cmd/mailchain/config/names"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/prompts"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/prompts"
 	"github.com/mailchain/mailchain/internal/pkg/encoding"
 	"github.com/spf13/cobra"
 )
