+++ cmd/mailchain/internal/commands/setup/receiver.go
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// nolint: dupl
 package setup
 
 import (
@@ -19,34 +20,32 @@ import (
 
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
 	"github.com/mailchain/mailchain/cmd/mailchain/config/names"
-	"github.com/mailchain/mailchain/internal/mailchain/commands/prompts"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/prompts"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper" // nolint: depguard
 )
 
-func Keystore(cmd *cobra.Command, keystoreType string) (string, error) {
-	keystoreType, err := selectKeystore(keystoreType)
+func Receiver(cmd *cobra.Command, chain, network, receiver string) (string, error) {
+	receiver, err := selectReceiver(chain, network, receiver)
 	if err != nil {
 		return "", err
 	}
-	if err := config.SetKeystore(cmd, keystoreType); err != nil {
+	if err := config.SetReceiver(viper.GetViper(), chain, network, receiver); err != nil {
 		return "", err
 	}
-
-	return keystoreType, nil
+	return receiver, nil
 }
 
-func selectKeystore(keystoreType string) (string, error) {
-	if keystoreType != names.Empty {
-		return keystoreType, nil
+func selectReceiver(chain, network, receiver string) (string, error) {
+	if receiver != names.Empty {
+		return receiver, nil
 	}
-	keystoreType, skipped, err := prompts.SelectItemSkipable(
-		"Key Store",
-		[]string{names.KeystoreNACLFilestore},
-		viper.GetString("storage.keys") != "")
+	receiver, skipped, err := prompts.SelectItemSkipable(
+		"Receiver",
+		[]string{names.EtherscanNoAuth, names.Etherscan},
+		viper.GetString(fmt.Sprintf("chains.%s.networks.%s.receiver", chain, network)) != "")
 	if err != nil || skipped {
 		return "", err
 	}
-	fmt.Printf("%q used for storing keys\n", keystoreType)
-	return keystoreType, nil
+	return receiver, nil
 }
